requirements.txt
bcrypt==4.0.0
capstone==5.0.0rc2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colored-traceback==0.3.0
cryptography==38.0.1
idna==3.3
intervaltree==3.1.0
Mako==1.2.2
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.11.0
pathlib2==2.3.7.post1
plumbum==1.7.2
psutil==5.9.2
pwntools==4.8.0
pycparser==2.21
pyelftools==0.29
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
ROPGadget==7.0
rpyc==5.2.3
six==1.16.0
sortedcontainers==2.4.0
unicorn==2.0.0
urllib3==1.26.12
